backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
aiosqlite==0.19.0
aioredis==2.0.1
cachetools==5.3.3
httpx==0.27.0
scrapy==2.11.1
pydantic==2.6.4
python-dotenv==1.0.1
loguru==0.7.2
beautifulsoup4==4.12.3
asyncio==3.4.3
typing-extensions==4.10.0
starlette==0.36.3
python-multipart==0.0.9
